refactor: move headset to VR directory
